refactor: reorganize code structure and improve navigation handling across screens

+++ app/src/main/java/com/example/project_enlishlearning/ui/vocabulary/EditVocabularyScreen.kt
@@ -33,7 +33,7 @@ fun EditVocabularyScreen(
 
     // 2. Tự động gọi ViewModel lấy dữ liệu khi mở màn hình
     LaunchedEffect(wordId) {
-        viewModel.loadWordById(wordId)
+        viewModel.loadWordById(wordId as Int)
     }
 
     var wordText by remember { mutableStateOf("") }
